+++ apps/io-profile-async/CHANGELOG.md
@@ -1,5 +1,18 @@
 # io-profile-async
 
+## 1.1.1
+
+### Patch Changes
+
+- 57dbe8c: Update ExpiredSessionAdvisor Queue Model (added expiredAt parameter).
+  Avoid sending E-Mail to user who do not have a validated E-Mail Address
+
+## 1.1.0
+
+### Minor Changes
+
+- 8f5f171: Enable dry-run mode for load testing
+
 ## 1.0.2
 
 ### Patch Changes